Move FusedQKV to slapo.op
