+++ examples/opt/model.py
@@ -12,8 +12,6 @@ from schedule import (
     remove_cast,
     replace_and_shard_mlp,
     replace_and_shard_attention,
-    replace_qkv,
-    shard_qkv,
     shard_word_embedding,
     trace_attention,
 )
